convert: Move GameVersion detection to module game_versions
