+++ openage/convert/driver.py
@@ -10,8 +10,7 @@ from subprocess import Popen, PIPE
 from tempfile import gettempdir
 
 from ..log import info, dbg
-
-from .main import GameVersion
+from .game_versions import GameVersion
 from .blendomatic import Blendomatic
 from .changelog import ASSET_VERSION, ASSET_VERSION_FILENAME
 from .colortable import ColorTable, PlayerColorTable
